+++ src/model/guild/partial_guild.rs
@@ -1468,7 +1468,7 @@ impl PartialGuild {
     #[cfg(all(feature = "cache", feature = "utils"))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u64 {
+    pub fn shard_id(&self, cache: impl AsRef<Cache>) -> u32 {
         self.id.shard_id(cache)
     }
 
@@ -1495,7 +1495,7 @@ impl PartialGuild {
     #[cfg(all(feature = "utils", not(feature = "cache")))]
     #[inline]
     #[must_use]
-    pub fn shard_id(&self, shard_count: u64) -> u64 {
+    pub fn shard_id(&self, shard_count: u32) -> u32 {
         self.id.shard_id(shard_count)
     }
